Passing axes to WrapPipe() instead of figure
This should prevent some of the lagginess with multiple generations in a session because Figure.add_subplot() is now being called less often.
The idea is that the same axes are reused for the pipe and wrapper plots with each new generation instead of trying to define new ones each time.
Generating output figures in new windows is also now possible as the default case when no axes are passed to PipeWrap(). There is a button for this.

Also, excel output is now not written if  the output filename is "" (which would mean the filename input box was clicked on but nothing was entered).